# ARROW-15336: [Go] Implement 'min_max_neon' with Arm64 GoLang Assembly