--- a/go/parquet/internal/utils/_lib/arch.h
+++ b/go/parquet/internal/utils/_lib/arch.h
@@ -22,6 +22,8 @@
     #define FULL_NAME(x) x##_sse4
 #elif __SSE3__ == 1
     #define FULL_NAME(x) x##_sse3
+#elif defined(__ARM_NEON) || defined(__ARM_NEON__)
+    #define FULL_NAME(x) x##_neon
 #else
     #define FULL_NAME(x) x##_x86
 #endif