--- a/go/parquet/internal/utils/_lib/min_max.c
+++ b/go/parquet/internal/utils/_lib/min_max.c
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <arch.h>
+#include "arch.h"
 #include <stdint.h>
 #include <limits.h>
 #include <math.h>