--- a/go/parquet/internal/utils/min_max_arm64.go
+++ b/go/parquet/internal/utils/min_max_arm64.go
@@ -18,9 +18,42 @@
 
 package utils
 
+import (
+	"os"
+	"strings"
+)
+import "golang.org/x/sys/cpu"
+
 func init() {
-	minmaxFuncs.i32 = int32MinMax
-	minmaxFuncs.ui32 = uint32MinMax
-	minmaxFuncs.i64 = int64MinMax
-	minmaxFuncs.ui64 = uint64MinMax
+    // Added ability to enable extension via environment:
+	// ARM_ENABLE_EXT=NEON go test
+	if ext, ok := os.LookupEnv("ARM_ENABLE_EXT"); ok {
+		exts := strings.Split(ext, ",")
+
+		for _, x := range exts {
+			switch x {
+			case "NEON":
+				cpu.ARM64.HasASIMD = true
+			case "AES":
+				cpu.ARM64.HasAES = true
+			case "PMULL":
+				cpu.ARM64.HasPMULL = true
+			default:
+				cpu.ARM64.HasASIMD = false
+				cpu.ARM64.HasAES = false
+				cpu.ARM64.HasPMULL = false
+			}
+		}
+	}
+	if cpu.ARM64.HasASIMD {
+		minmaxFuncs.i32 = int32MaxMinNEON
+		minmaxFuncs.ui32 = uint32MaxMinNEON
+		minmaxFuncs.i64 = int64MaxMinNEON
+		minmaxFuncs.ui64 = uint64MaxMinNEON
+	} else {
+		minmaxFuncs.i32 = int32MinMax
+		minmaxFuncs.ui32 = uint32MinMax
+		minmaxFuncs.i64 = int64MinMax
+		minmaxFuncs.ui64 = uint64MinMax
+	}
 }