@guyuqi 2022-01-17
Test in `arrow/go/parquet/internal/encoding`
ARM_ENABLE_EXT=NEON go test
```
PASS
ok      github.com/apache/arrow/go/v7/parquet/internal/encoding 9.788s
```