@guyuqi 2022-01-17
@zeroshade Added min_max_neon for parquet encoding. 
Get 4x ~ 6x performance uplift in BenchmarkPlainEncodingInt32/Int64/Float32/Float64.